Prevent the pause menu from creating any threads.
In Singleplayer and GTAO, certain pause menu tabs may stream in and start additional scripts/threads to fill in dynamic content. For example to handle clicking on race/deathmatch join blips, or fill in the mission objective in the GAME tab.

However this is somehow causing the crashes described in #3079, as the pause menu attempts to start a non-existent program. Not sure how that's happening but since this functionality is completely unused in FiveM, might as well patch it out.